+++ src/main/kotlin/org/hugopalma/domaintranslator/settings/SettingsComponent.kt
@@ -1,5 +1,6 @@
 package org.hugopalma.domaintranslator.settings
 
+import com.intellij.openapi.ui.ComboBox
 import com.intellij.ui.components.JBCheckBox
 import com.intellij.ui.components.JBLabel
 import com.intellij.ui.components.JBTextField
@@ -8,6 +9,8 @@ import com.intellij.util.ui.JBFont
 import com.intellij.util.ui.JBUI
 import com.intellij.util.ui.UIUtil
 import okio.Path
+import org.jetbrains.kotlin.idea.base.util.onTextChange
+import java.util.*
 import javax.swing.JComponent
 import javax.swing.JPanel
 import javax.swing.plaf.LabelUI
@@ -18,14 +21,26 @@ class SettingsComponent {
     private val showInlays = JBCheckBox("Display inlays", true)
     private val dictionaryFile = JBTextField()
     private val hideInlays = JBTextField()
+    private val useIntellijLanguage = JBCheckBox("Use system language", true)
+    private var languagesPanel: JPanel
+    private val languages = ComboBox(getAllLanguages().toTypedArray())
 
     init {
         hideInlaysPanel = FormBuilder.createFormBuilder()
             .addLabeledComponent("Words to hide in inlay:", hideInlays, 1, false)
             .addComponent(CommentLabel("List of words that won't be translated in inlay separated by comma"))
             .panel
+        showInlays.addChangeListener { hideInlays.isEnabled = showInlays.isSelected }
 
-        showInlays.addChangeListener { hideInlaysPanel.isVisible = showInlays.isSelected }
+        languagesPanel = FormBuilder.createFormBuilder()
+            .addLabeledComponent("Show translation in language", languages, 1, false)
+            .panel
+        useIntellijLanguage.addChangeListener { languages.isEnabled = !useIntellijLanguage.isSelected }
+
+        dictionaryFile.onTextChange { event ->
+            useIntellijLanguage.isEnabled = dictionaryFile.text.isEmpty()
+            languages.isEnabled = dictionaryFile.text.isEmpty()
+        }
 
         mainPanel = FormBuilder.createFormBuilder()
             .addLabeledComponent("Dictionary file:", dictionaryFile, 1, false)
@@ -34,8 +49,18 @@ class SettingsComponent {
             .addComponent(showInlays, 1)
             .addComponent(CommentLabel("Should the inline domain translation be displayed"))
             .addComponentToRightColumn(hideInlaysPanel)
+            .addComponent(useIntellijLanguage)
+            .addComponentToRightColumn(languagesPanel)
             .addComponentFillVertically(JPanel(), 0)
             .panel
+    }
+
+    private fun getAllLanguages(): List<Language> {
+        return Locale.getAvailableLocales()
+            .filter { locale -> locale.language.isNotEmpty() }
+            .distinctBy { locale -> locale.displayLanguage }
+            .sortedBy { locale -> locale.displayLanguage }
+            .map { locale -> Language(locale) }
     }
 
     fun getPanel(): JPanel? {
@@ -60,7 +85,7 @@ class SettingsComponent {
 
     fun setShowInlays(value: Boolean) {
         showInlays.isSelected = value
-        hideInlaysPanel.isVisible = value
+        hideInlaysPanel.isEnabled = value
     }
 
     fun getHideInlays(): String {
@@ -69,6 +94,23 @@ class SettingsComponent {
 
     fun setHideInlays(value: String) {
         hideInlays.text = value
+    }
+
+    fun setUseSystemLanguage(value: Boolean) {
+        useIntellijLanguage.isSelected = value
+        languages.isEnabled = !value
+    }
+
+    fun getUseSystemLanguage(): Boolean {
+        return useIntellijLanguage.isSelected
+    }
+
+    fun getSelectedLanguage(): Language? {
+        return languages.selectedItem as Language?
+    }
+
+    fun setSelectedLanguage(language: Language?) {
+        languages.selectedItem = language
     }
 }
 
@@ -80,5 +122,22 @@ class CommentLabel(text: String) : JBLabel(text) {
     override fun setUI(ui: LabelUI) {
         super.setUI(ui)
         font = JBFont.medium()
+    }
+}
+
+class Language(val locale: Locale) {
+    override fun toString(): String {
+        return locale.displayLanguage + " (${locale.language})"
+    }
+
+    override fun equals(other: Any?): Boolean {
+        if (this === other) return true
+        if (javaClass != other?.javaClass) return false
+
+        return locale.language == (other as Language).locale.language
+    }
+
+    override fun hashCode(): Int {
+        return locale.hashCode()
     }
 }